+++ tests/_tests/string_magic_tests.R
@@ -845,7 +845,11 @@ test(txt, "The max value is .")
 
 # nesting
 x = c(15, 550)
-txt = string_magic("The values are{& length(x) < 5 ; ': {C ? x}' ; 'too many'}.")
+txt = string_magic("The values are{& length(x) < 5 ; ': {C ? .}' ; ' too many'}.")
+test(txt, "The values are: 15 and 550.")
+
+data = data.frame(x = x)
+txt = string_magic("The values are{& length(data$x) < 5 ; ': {C ? .}' ; ' too many'}.")
 test(txt, "The values are: 15 and 550.")
 
 # special if else